pyproject.toml
[tool.poetry]
name = "deep-studio"
version = "0.1.0"
description = ""
authors = ["sangwon lee <leeh8911@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
pytest = "^8.3.4"
sphinx = "^8.1.3"
numpy = "^2.1.3"
flake8 = "^7.1.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
